Stubbed out logic to load the username and password from the settings file if an <id> element is provided and username and password are both null.
Added code to get the username and password from the settings.xml file
if an id is provided (with null username and password values).

Added libraries required to process the contents of the settings file as
well as the libraries used to test maven plug-ins.

Copied a stubbed out example pom.xml and settings.xml file into the
project.

Stubbed out a base TestCase with some helper methods and added an empty
test class for the AbstractServerConnection (Mojo).

Eliminated HTTP imports that weren't provided by Maven dependencies (or
needed by the plug-in)

Fixed the annotation on the new settings parameter.

Cleaned up the utility methods and added comments.

Added the rest of the pom.xml files and settings.xml file needed to test
how credentials are acquired by the plugin.

Added the rest of the settings.xml and pom.xml files needed to
completely test the plugin's acquisition of credentials.

Added Mockito to the test dependencies so that the plugin's log could be
mocked and analyzed.

Completed the logic that determines how credentials are acquired by the
plugin.

Finished writing tests for the 6 usable permutations of how credentials
can be acquired by the plugin.

Changed the last few log messages related to the acquisition of
credentials from info to debug level.

Added version properties for all dependencies.

Update copyright headers on tests.